% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRBplots-package.R
\docType{package}
\name{WRBplots-package}
\alias{WRBplots}
\alias{WRBplots-package}
\title{WRBplots: Make pretty, standard plots for a variety of WRB uses}
\description{
Functions that facilitate the generation of custom plots from data downloaded from Aquarius.
}
\author{
\strong{Maintainer}: Cole Fischer \email{cole.fischer@hotmail.com} (\href{https://orcid.org/0000-0002-7266-5760}{ORCID})

}
\keyword{internal}
