% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YOWNplot_1YrStat.R
\name{YOWNplot_1YrStat}
\alias{YOWNplot_1YrStat}
\title{YOWN 1-Year Stat Plot Generation Function}
\usage{
YOWNplot_1YrStat(
  AQID,
  timeSeriesID = "Wlevel_bgs.Calculated",
  saveTo = "desktop",
  AQTSServerID = "https://yukon.aquaticinformatics.net/AQUARIUS",
  login = Sys.getenv(c("AQUSER", "AQPASS"))
)
}
\arguments{
\item{AQID}{Identity of YOWN site in the following format: "YOWN-XXXX" or "YOWN-XXXXD"}

\item{timeSeriesID}{Identity of the time series exactly as written in Aquarius (eg."Wlevel_bgs.Calculated")}

\item{saveTo}{Location for data files to be saved. Default is creation of a new folder on your desktop.}

\item{AQTSServerID}{Defaults to Yukon Water Resources Branch Aquarius web server}

\item{login}{Your Aquarius login credentials as a character vector of two (eg. c("cmfische", "password") Default pulls information from your .renviron profile; see details.}
}
\value{
Writes .pdf of a 1 year water level plot for a YOWN site
}
\description{
This function checks length of YOWN site record, if greater than 2 years calculates and plots most recent year's level data and historical statistics. If period of record is shorter than 2 years, outputs most recent year of level data only.
}
\details{
To store login credentials in your .renviron profile, call usethis::edit_r_environ() and enter your username and password as value pairs, as AQUSER="your username" and AQPASS = "your password".
}
