% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YOWNplot_SiteCompare.R
\name{YOWNplot_SiteCompare}
\alias{YOWNplot_SiteCompare}
\title{YOWN site comparative plot generation, in m bgs}
\usage{
YOWNplot_SiteCompare(
  YOWNindex,
  tsunit,
  AQTSServerID = "https://yukon.aquaticinformatics.net/AQUARIUS",
  chartRange = "all",
  chartXInterval = "1 month",
  saveTo = "desktop",
  login = Sys.getenv(c("AQUSER", "AQPASS"))
)
}
\arguments{
\item{YOWNindex}{Character vector of YOWN site IDs (eg. c("YOWN-2201S", "YOWN-2201D", "YOWN-2202"))}

\item{tsunit}{Desired timeseries for plotting, CHOOSE FROM: "btoc", "bgs", "asl"}

\item{AQTSServerID}{Defaults to Yukon Water Resources Branch Aquarius web server}

\item{saveTo}{Location for data files to be saved. Default is publication to a new folder on your desktop.}

\item{login}{Your Aquarius login credentials as a character vector of two (eg. c("cmfische", "password") Default pulls information from your .renviron profile; see details.}

\item{WSCindex}{Character vector of WSC site IDs (eg. c("09AB004", "09AH001"))}
}
\value{
Writes .pdf plot of WSC and YOWN data
}
\description{
Plots multiple YOWN stations on the same chart
}
\details{
To store login credentials in your .renviron profile, call usethis::edit_r_environ() and enter your username and password as value pairs, as AQUSER="your username" and AQPASS = "your password".
Marsh Lake: 09AB004
Yukon River at Carmacks: 09AH001
}
