% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YOWNplot_WSC.R
\name{YOWNplot_WSC}
\alias{YOWNplot_WSC}
\title{YOWN-WSC comparative plot generation relative to a common datum}
\usage{
YOWNplot_WSC(
  YOWNindex = c("YOWN-2201S", "YOWN-2201D", "YOWN-2202", "YOWN-2203", "YOWN-2204",
    "YOWN-2205"),
  WSCindex = c("09AH001"),
  saveTo = "desktop",
  title = "YOWN wells vs. WSC Sites",
  chartXinterval = "1 month"
)
}
\arguments{
\item{YOWNindex}{Character vector of YOWN site IDs (eg. c("YOWN-2201S", "YOWN-2201D", "YOWN-2202"))}

\item{WSCindex}{Character vector of WSC site IDs (eg. c("09AB004", "09AH001"))}

\item{saveTo}{Location for data files to be saved. Default is publication to a new folder on your desktop.}

\item{AQTSServerID}{Defaults to Yukon Water Resources Branch Aquarius web server}

\item{login}{Your Aquarius login credentials as a character vector of two (eg. c("cmfische", "password") Default pulls information from your .renviron profile; see details.}
}
\value{
Writes .pdf plot of WSC and YOWN data
}
\description{
Plots YOWN data alongside WSC stations, in m asl. Only works for stations that have values in the "Elevation" field in their Aquarius information page.
}
\details{
To store login credentials in your .renviron profile, call usethis::edit_r_environ() and enter your username and password as value pairs, as AQUSER="your username" and AQPASS = "your password".
Marsh Lake: 09AB004
Yukon River at Carmacks: 09AH001
}
