% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRBplots-package.R
\docType{package}
\name{WRBplots-package}
\alias{WRBplots}
\alias{WRBplots-package}
\title{WRBplots: Make pretty, standardized plots for a variety of WRB uses}
\description{
Functions that facilitate the generation of custom plots from several different data sources in a consistent format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YukonWRB/WRBplots}
  \item Report bugs at \url{https://github.com/YukonWRB/WRBplots/issues}
}

}
\author{
\strong{Maintainer}: Cole Fischer \email{cole.fischer@hotmail.com} (\href{https://orcid.org/0000-0002-7266-5760}{ORCID})

Authors:
\itemize{
  \item Ghislain de Laplante \email{ghislaindel@gmail.com} (\href{https://orcid.org/0000-0002-5093-9185}{ORCID})
}

}
\keyword{internal}
