% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRBplots-package.R
\docType{package}
\name{WRBplots-package}
\alias{WRBplots}
\alias{WRBplots-package}
\title{WRBplots: Make pretty, standard plots for a variety of WRB uses}
\description{
Functions that facilitate the generation of custom plots from several different data sources. Currently set up to work with data directly from Aquarius and from the local hydrometric database.
}
\keyword{internal}
