% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YOWNplot.R
\name{YOWNplot}
\alias{YOWNplot}
\title{YOWNplot}
\usage{
YOWNplot(
  AQID = "YOWN-1925",
  timeSeriesID = "Wlevel_bgs.Calculated",
  chartXinterval = "auto",
  chartXlimits = "all",
  stats = TRUE,
  smooth = TRUE,
  saveTo = "desktop",
  login = Sys.getenv(c("AQUSER", "AQPASS")),
  AQTSServerID = "https://yukon.aquaticinformatics.net/AQUARIUS"
)
}
\arguments{
\item{AQID}{YOWN location for which a plot will be generated}

\item{timeSeriesID}{Aquarius time series ID exactly as in Aquarius (ie. "Wlevel_bgs.Calculated"). Defaults to m bgs.}

\item{chartXinterval}{X axis interval, can be specified "auto" for best fit calculation, or as desired (ie. "1 day", "1 month", "1 year", etc.). Defaults to "auto"}

\item{chartXlimits}{X axis limits, can be "all" for all data, "1yr" for most recent year of data, or vector of 2 in format c("2020/01/01 00:00:00", "2023/01/01 00:00:00"). Defaults to "all"}

\item{stats}{TRUE/FALSE Add historical max/min and average to plot, where sufficient data exists.}

\item{smooth}{TRUE/FALSE Plot rolling 5 day average instead of individual values.}

\item{saveTo}{Directory in which the plot will be saved. Can specify "desktop" to create YOWN ID folder on desktop as save directory.}

\item{login}{Aquarius username and password, taken from Renviron file}

\item{AQTSServerID}{Aquarius server ID}
}
\value{
Writes a .pdf containing YOWN data in the specified directory.
}
\description{
Generalized YOWN plotting function
}
