% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YOWNplot.R
\name{YOWNplot}
\alias{YOWNplot}
\title{Graphing utility for YOWN wells}
\usage{
YOWNplot(
  AQTSServerID = "https://yukon.aquaticinformatics.net/AQUARIUS",
  AQTSUsername = "gtdelapl",
  AQTSPassword = "WQ*2021!",
  dateRange = "all",
  timeRange = c("00:00:00", "23:59:59"),
  AQID = "YOWN-1907",
  timeSeriesID = "Water Temp.TEMPERATURE",
  chartXInterval = "1 year",
  chartType = "Temperature",
  saveTo = "C:/Users/gtdelapl/Desktop",
  specName = NULL
)
}
\arguments{
\item{AQTSServerID}{The web address of your Aquarius server.}

\item{AQTSUsername}{The Aquarius username you wish to use.}

\item{AQTSPassword}{The password associated with the username.}

\item{dateRange}{The date range your wish to graph for. Use "all" for the entirety of the time-series (1950-01-01 to today) or specify two dates as such: c("2000-01-01", "2021-01-01")}

\item{timeRange}{Defaults to the entire day. Set in the same manner as dateRange with hh:mm:ss format if you wish to plot only a portion of a day; in that case dateRange should have two identical dates.}

\item{AQID}{The Aquarius ID of the site you wish to graph.}

\item{timeSeriesID}{The time-series you wish to graph, matching exactly how it is written in Aquarius (case sensitive).}

\item{chartXInterval}{The x-axis interval markings, specified as per scales::date_breaks}

\item{chartType}{The type of chart. Currently supporting only "Level", "SpC", "Temperature".}

\item{saveTo}{The directory where you wish to save your graph. Will create a folder with the AQID if it does not yet exist and saves the image as a .png with a name composed of the chartType, AQID and the current date.}

\item{specName}{Specify a file name here only if you wish to override the default naming of AQID and the current date; file name will still start with chartType. You should modify the NULL default if generating two graphs that will be named identically (same CharType, AQID, and date but different )}
}
\value{
A .png image of the time series with Yukon logos, saved to the directory specified in saveTo.
}
\description{
Create standardized graphs for YOWN wells from data hosted on the Aquarius server. Each graph contains information specific to the site and the time-series requested, and is complemented by the YG and water logos in the upper corners. Saves the graph to the path specified by the user.
}
